refactor: Meal.tsx의 초기화 로직을 initializeData.ts 로 이동
