+++ constants.ts
@@ -2,4 +2,5 @@ export const FAVORITE_STORAGE_KEY = {
   cafe: 'favoriteCafeList',
   mart: 'favoriteMartList',
   shuttle: 'favoriteShuttleList',
+  meal: 'favoriteMealList',
 };
